requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
